feat: 자주 사용하는 Composable 메서드 구현
ComposableUtil.kt 내부 코드를 StateDispatcher 파일 생성하여 이관함.

Related To: #6
