+++ client/presentation/src/main/java/com/everyone/movemove_android/base/StateDispatcher.kt
@@ -1,0 +1,29 @@
+package com.everyone.movemove_android.base
+
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import kotlinx.coroutines.flow.SharedFlow
+
+@Composable
+inline fun <reified STATE, EVENT, EFFECT> use(
+    viewModel: BaseContract<STATE, EVENT, EFFECT>
+): StateDispatchEffect<STATE, EVENT, EFFECT> {
+    val state by viewModel.state.collectAsStateWithLifecycle()
+
+    val dispatch: (EVENT) -> Unit = { event ->
+        viewModel.event(event)
+    }
+
+    return StateDispatchEffect(
+        state = state,
+        effectFlow = viewModel.effect,
+        dispatch = dispatch,
+    )
+}
+
+data class StateDispatchEffect<STATE, EVENT, EFFECT>(
+    val state: STATE,
+    val dispatch: (EVENT) -> Unit,
+    val effectFlow: SharedFlow<EFFECT>,
+)
